refactor: 파일 이동 및 JwtConfig.kt 제거
