+++ src/main/kotlin/com/yapp/demo/auth/external/kakao/feign/KakaoUserInfoFeignClient.kt
@@ -1,6 +1,6 @@
 package com.yapp.demo.auth.external.kakao.feign
 
-import com.yapp.demo.auth.external.kakao.response.KakaoUserInfoResponse
+import com.yapp.demo.auth.external.kakao.feign.response.KakaoUserInfoResponse
 import org.springframework.cloud.openfeign.FeignClient
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestHeader
