--- src/main/kotlin/com/yapp/demo/common/config/JwtConfig.kt
@@ -1,9 +1,0 @@
-package com.yapp.demo.common.config
-
-import com.yapp.demo.common.config.properties.JwtProperties
-import org.springframework.boot.context.properties.EnableConfigurationProperties
-import org.springframework.context.annotation.Configuration
-
-@Configuration
-@EnableConfigurationProperties(JwtProperties::class)
-class JwtConfig
