+++ src/main/kotlin/com/yapp/demo/common/config/properties/JwtProperties.kt
@@ -1,10 +1,12 @@
 package com.yapp.demo.common.config.properties
 
 import org.springframework.boot.context.properties.ConfigurationProperties
+import org.springframework.stereotype.Component
 
+@Component
 @ConfigurationProperties(prefix = "jwt")
 data class JwtProperties(
-    val secret: String,
-    val accessTokenExpiryTime: Long,
-    val refreshTokenExpiryTime: Long,
+    var secret: String = "",
+    var accessTokenExpiryTime: Long = 0L,
+    var refreshTokenExpiryTime: Long = 0L,
 )
